Criação do Post no TransactionController ele valida se o Client existe e está ativo, depois ele salva a Transaction no banco de dados e depois ele manda uma mensagem pro KafkaTransactionProducer transformando o Transaction em String para o Kafka conseguir serializar...

+++ service-client/src/main/java/com/nttdata/application/usecase/client/DeleteClient.java
@@ -1,6 +1,7 @@
 package com.nttdata.application.usecase.client;
 
 import com.nttdata.application.repository.ClientRepository;
+import com.nttdata.domain.exceptions.ClientInactiveException;
 
 public class DeleteClient {
     private final ClientRepository repository;
@@ -10,7 +11,12 @@ public class DeleteClient {
     }
 
     public void deleteClient(Long id) {
-        repository.deleteClient(id);
+        boolean active = repository.verifyActive(id);
+        if (active) {
+            repository.deleteClient(id);
+        }else {
+            throw new ClientInactiveException("Client already inactive!");
+        }
     }
 
 }
